Don't register empty code fragments
Native dynlink registers a code fragment if both pointers are
successfully loaded. However, if _shared_startup has no code then the
start address of this code fragment will be the same as the start
address of the actual unit being dynlinked, creating a duplicate
insertion. Prior to 5.00 this was simply erased by the next insertion,
but the issue surfaced as the new skiplist implementation didn't replace
the pointer on the second insertion.