Fix race condition on signal mask access in `domain_thread_func`
Currently, the `pthread_sigmask(SIG_SETMASK, &p->mask, NULL)` in
`domain_thread_func` takes place after the handshake with the parent
`caml_domain_spawn` is completed, at a time where `p` (which points
into the stack frame of `caml_domain_spawn) is no longer safe to access.

(This is actually noted in a helpful comment, `/* Cannot access p below here. */`).

The fix is to take a local copy of the signal mask early in
`domain_thread_func`, before the handshake is completed, at a time
when the parent is blocked.

I seized the opportunity to pass the signal mask (through the `struct
domain_startup_params`) by reference instead of by value, saving one
copy (!) and making the struct smaller.