Reimplement Thread.exit using an exception
Previously, Thread.exit tried to behave exactly like pthread_exit():
it stops the current thread immediately, but doesn't stop the whole
program as long as other threads keep running.

As discussed in #9071, this behavior is a problem for resource management
("finally" functions are not called).

As discussed during the Multicore merge review and reported in #10927,
this behavior is a problem in the presence of multiple domains.

As first proposed in #9100, this PR reimplements Thread.exit by
simply raising a dedicated Thread.Exit exception.  This is good for
resource management ("finally" functions are properly called)
and simplifies the implementation tremendously.

It does introduce incompatibilities with OCaml 4.  In particular, if
`Thread.exit` is called from the main program or from a domain but not
from a thread started by `Thread.create`, the exception `Thread.Exit`
is not automatically caught and will abort the program / the domain,
unless explicitly handled by the program.