Restore some comments & polling locations
Polling locations were tightened for memprof. They partially
disappeared during multicore's rebase on top of 4.12 since multicore
did not support memprof or the new polling machinery. In addition, it
was inconsistent whether these changes and/or the comments relating
them to memprof were kept or removed.

These changes will be easy to miss when working to restore memprof
unless someone goes through the diff of the multicore PR. Since I have
already done it, this will make sure that nothing is accidentally
lost.

This commit is meant to be behaviour-preserving.

No change entry needed.