Make <caml/sync.h> more abstract and refactor implementation of sync.c
<caml/sync.h> declares only the high-level, platform-agnostic mutex
and condition variable operations that operate on OCaml values.

domain.c is adjusted to use these high-level operations to work with
termination mutexes.

The POSIX implementation of mutexes and condition variables that used
to be in <caml/sync.h> is moved to sync_posix.h (an unexported header,
included in sync.c only) and extended with all POSIX-specific
operations.

sync.c and otherlibs/systhreads/st_posix.h are adjusted accordingly.