Ocamldoc: add multicore tags
This commit adds basic support for the following tags intended to be used
to document concurrency-safety in OCaml 5:

@domain-safe
@systhread-safe
@fiber-safe
@concurrent-unsafe

or equivalently

@concurrency {domain-safe,systhread-safe,fiber-safe,concurrent-unsafe}

Those new tags are currently only displayed in the html backend,
while waiting for discussion on the design.

In particular, currently a default @concurrent-unsafe tag is displayed
for module and module types in there is no explicit concurrency tag.
We probably don't want to do that on individual items.