Minimize changes to domain.c
This commit reverts most of the changes to domain.c, so that
the termination lock is released late and manipulated as a
pointer outside the heap, like before.  It looks safer this way.

To support this, <caml/sync.h> is changed once more to export
the type `sync_mutex` and two lock / unlock functions defined in sync.c
and usable in domain.c.