Fix caching of stacks when multiple domains running
Each domain keeps a cache of allocated stacks that are reused
when new fibers are created or stacks are resized. They are
stored in an array of 'buckets' (linked lists) indexed by size,
and the specific bucket in which they belong is stored in the
stack structure so they can be put back quickly when finished.
Before this patch the stack referenced the bucket via an
explicit pointer initialised when the stack was first
allocated, which causes problems when the stack is being
cleaned up by a different domain to the one in which it was
created, as we are then manipulating a data structure that
should exclusively belong to another domain.

This patch removes the pointer to the linked list and replaces
it with the index into the array of linked lists instead. With
this the domain cleaning up the stack can put it back in its
own cache of stacks rather than trying to give it back to the
domain that allocated it in the first place.

Co-authored-by: KC Sivaramakrishnan <kc@kcsrk.info>
Co-authored-by: Tom Kelly <ctk21@cl.cam.ac.uk>
Fixes: #11094