win32unix `create_process`: duplicate std handles in `create_process` iff they're not inheritable  (#10807)
* win32unix: test if create_process leaks std handles
* win32unix: duplicate handle in create_process iff not inheritable

The current code makes the assumption that handles are not inheritable
(close-on-exec), and duplicates handles to mark them inheritable
(keep-on-exec) for the child process. That assumption is only
mentioned in #650, and not documented in the manual.

It's possible however that the calling code gives to create_process
handles that are already inheritable: either because the programmer
isn't aware of the security implications, or because the programmer
expects create_process to be a "simple" wrapper around WinAPI's
CreateProcess; in which case setting the handle inheritable would be
the responsibility of the calling code.

Converting the canonical example from Microsoft documentation [1]
without looking at win32unix's create_process would lead to this
problem.

In either case, the current create_process would needlessly duplicate
an already inheritable handle and the child process would inherit both
the original and the duplicated handles.

Consider the following pipeline: `writer | child | reader`.
The problem with the child process having duplicated "leaked" stdin
and stdout handles is that while it is running, it cannot signal to
the writer that it should stop writing: the writer will not see a
broken pipe as the child still has the leaked duplicated handle
open. The same thing happens to the reader with the stdout of the
child. Usually stdin/stdout are bound to the lifetime of the process,
so this is not too much of a problem, but if the child itself spawns
processes that should outlive it, the whole thing can become messy.

We can instead check whether the given handles are inheritable, and
only duplicate them if they're not. The calling code is still
responsible for closing the original handle.

[1]: https://docs.microsoft.com/en-us/windows/win32/procthread/creating-a-child-process-with-redirected-input-and-output