Add C function `caml_thread_has_lock`
This function can be called from any C thread and returns true if the
thread belongs to a domain and holds the lock of its domain.

It necessary for implementing C functions which are generic on whether
the domain lock is held or not. In case `caml_thread_has_lock` returns
false, the programmer can for instance:
- report an error
- try to acquire the domain lock
- do something else (e.g. delay an operation until the lock is
  acquired)

One application is for the OCaml-Rust FFI: in some situations, it is
not possible to enforce with static typing that some function runs
only when the runtime lock is held (notably destructors). This
function lets us implement:
- a dynamic check of whether the lock is held (which plays well with
  the lifetime-based GC safety),
- a way to acquire the domain lock only if it is not already held.

In OCaml 4, it is possible to implement this feature on the
programmer-side (using internals) by customizing the enter/leave
runtime hooks. Alas, this strategy does not work in OCaml 5. One would
have to somehow install hooks after systhread is loaded, but before
the second domain is spawned (after which hooks can no longer be
modified), and this is very hard to do. So we need a native solution
for OCaml 5.