Windows: replace `socket` with `WSASocket` and set `WSA_FLAG_NO_HANDLE_INHERIT` (#10809)
* win32unix: replace calls to socket with WSASocket

The call to

    socket(domain, type, protocol);

is equivalent to

    WSASocket(domain, type, protocol, NULL, 0, WSA_FLAG_OVERLAPPED);

the latter offering more (needed) control over the creation of
sockets.

* Use WSA_FLAG_NO_HANDLE_INHERIT on Windows for cloexec sockets

On Windows, creating a socket then using `SetHandleInformation` on it
to make it non-inheritable (close-on-exec) is prone to a race
condition: the socket isn't created atomically non-inheritable and a
child process could inherit the socket between its creation and the
call to `SetHandleInformation`. This can lead to a deadlock of the
parent process, as explained in [KB2398202].

The hotfix [KB2398202] introduced the `WSA_FLAG_NO_HANDLE_INHERIT` to
be used with `WSASocket` to create non-inheritable sockets
atomically. It serves the same purpose as `SOCK_CLOEXEC` for
`socket(2)` on Unix systems.

This flag is supported on Windows 7 with SP1, Windows Server 2008 R2
with SP1, and later. On older systems, if `WSASocket` fails with
`WSAEINVAL` when using the flag, we can fallback to
`SetHandleInformation`.

Wrap `WSASocket` in new `win32_socket` to share the common inheritance
code.

[KB2398202]: https://support.microsoft.com/en-us/topic/an-application-may-stop-responding-when-the-application-closes-a-socket-connection-or-shuts-down-41321a1f-d80c-6975-98df-ae499d63133c