Fix parallel rebuild of dynlink.cma
GNU make needs to know that the compiler libs .cmo and .cmi are produced
simultaneously by the same recipe. The fix exploits the fact that while
GNU make multi-target rules are quite new (GNU make 4.3+), multi-target
pattern rules have existed for a long time.

Previously, only the dependency of byte/dynlink_compilerlibs.cmi on
byte/dynlink_compilerlibs.cmo was given to GNU make. GNU make
(correctly) infers that as no rules emit byte/dynlink_compilerlibs.cmi,
it can't change during the build. It therefore only checks the state of
byte/dynlink_compilerlibs.cmo at the start of the build, so if it
then changes _during_ the build, it does not propagate any rebuilds to
things which depended on byte/dynlink_compilerlibs.cmi which in this
case includes dynlink_common.cmo. The lack of this dependency allowed
dynlink.cma to be relinked with inconsistent assumptions when rebuilding
in parallel.