ocamldep: add -strict, error if input doesnt exist
Closes #8625
See ocaml-flambda/flambda-backend#892

When a nonexistent input file is passed to ocamldep, it is ignored and
the error is not reported. This PR adds a `-strict` flag that prevents
this behavior.

Consider the following situation:
- on one side, ocamldep is being invoked as `ocamldep -modules a.ml`
- on the other side, `a.ml` is unlinked and created again.

(this corresponds to what happens when `dune` is running and `hg` or
`git` is updating its tree in the meantime - updates are not always
atomic if hardlinks are involved for example)

We want to make sure that when ocamldep exits with 0, its output
corresponds to either the old or the new version of the file. This is
enabled by `-strict`.