Bump default maximum stack size (#11238)
Now that stack allocation is entirely under the control of the OCaml
runtime system, it's time to use more generous stack limits than
those provided by operating systems.

This commit sets the default maximum stack size to 128 Miwords,
i.e. 1 Gib for 64-bit platforms and 512 Mib for 32-bit platforms.

* tests/tool-ocaml/t360-stacks-2.ml: use a fixed max stack size

So that the outcome of the test doesn't depend on the default max stack size.