Add tests for Domain.recommended_count
I've added the stub just to fetch Max_domains which is not exposed, I also
defined CAML_INTERNAL, I hope this is ok on tests.

The test is very defensive, we try external tools that should return a number
that is atleast larger than the one we computed, instead of doing a bunch of OS
dependent stuff, just try them all:

 * nproc should work on linux
 * sysctl hw.ncpu should work on all BSDs including Darwin, hw.ncpu is the old
school "big number", so it's the largest possible value (as in doesn't take into
account affinity or online cpus).

In case no tools are available (no nproc, no sysctl hw.ncpu) assume the test
passed (still check max and min though), we can't infer the environment we're
being built, it might be chrooted and etcetera.

The split into two tests is so that we can at least test the very basic stuff on
windows.