optimize roots-scanning for the bytecode stack
The bytecode stack may contain naked code pointers. In general,
when scanning the stack to call scanning actions, we must check
the code-fragment table to filter out those naked pointers, but this
is fairly expensive.

In the specific case of the minor gc, we are scanning the stack with
[oldify_one], which gracefully accepts naked pointers (it ignores
all pointers outside the minor heap).

The current PR passes "scanning action flags" to all root scanners
(including user-provided hooks), to track whether their scanning
actions supports naked pointers, and disable the code-fragment check
in this case.

This fixes a 50% performance slowdown of the bytecode interpreter
compared to OCaml 4 on some benchmarks with large stacks full of
non-immediate values, such as the stacks.ml benchmark of
Sandmark (in 'floats-large' mode).