Make the documentation of `string_of_float` more precise
This changes the wording since the output of the function is not "the"
string representation of a float but rather "a" string representation.
In fact a very specific one. I would say it is safe to state that
`string_of_float` will always return a decimal notation, but the
precision of it is probably an implementation detail.