ocamlmktop: add a new initialization module
This initialization module is linked before any of the user-provided
modules and reads "+compiler-libs/topdirs.cmi". This step is required
before installing any custom printers.
Custom repl build before OCaml 5 are unaware of this requirement and thus
ocamlmktop needs to initialize this part of the toplevel environment itself
to preserve backward compatibility.