Tweak the Check Changes job
At present, if a PR is opened with no Changes entry, the job fails. If
the no-changes-entry-needed label is then added, the job is skipped.
However, when the PR is merged, GitHub then reports a failed check.

This restores the old behaviour where we use the GitHub API to check the
labels on the PR but, unlike on Travis, uses the GITHUB_TOKEN set-up
within the runner so that we (hopefully) don't upset GitHub API rate
limiting.