Do not remove compiler-generated files if they are not regular files
An example of thing that can go wrong is if we're root and we do
ocamlc -o /dev/null.

In all uses of `Misc.remove_file` in the compiler, the file being
removed is intended to be a regular file.  If it is not, something
unexpected is going on and it's safer not to remove.

Fixes: #11302.