Handle effect names in Effect.Unhandled exception stringify
It seemed sensible not to duplicate the logic to turn sum types into
strings that is currently in use for exceptions. However, since it's
not exposed in Printexc, and the exception being matched is in
effect.ml, which depends on Printexc, I've opted for matching the
string "Stdlib.Effect.Unhandled" and assuming the single argument is
the effect.

(cherry picked from commit 9f4691d2ace36bc803c41a9c39f36b11441be12c)