Add syntax for generative functor application
Previously, writing [F ()] was the same as writing
[F (struct end)], even though the latter looks like the
use of an applicative functor, not a generative one.

This commit, originally written by Frédéric Bour <fred@tarides.com>,
adds new syntax to our AST to represent generative functor
application and propagates this change throughout the compiler.

In addition, it adds a new warning, 73, to report when a user
has written [F (struct end)] but should now update to [F ()].