Build system: introduce a general framework to specify compiler flags
This commit introduces a set of build variables that will be used to
define which compiler flags to use to build the compiler.

At this stage, the framework is used only in the root Makefile
so it does only define the variables required by this root Makefile.

The framework will be extended as required to take into account
the different sub-makefiles as they get merged into the root one.

The build variables are defined in Makefile.build_config.in rather than in
Makefile.common so that the flags can, if necessary, be controlled
by the configure script (see for instance OC_NATIVE_CFLAGS).