Avoid data race in method caching in ocamlrun (#11512)
When calling a method with the opcode GETPUBMET, the method address is
resolved and possibly cached by modifying a word in the bytecode. If
multiple domains execute the same GETPUBMET opcode concurrently, this is
technically a data race. However, any integer value in the correct range
is a safe value. Therefore we inform the compiler of potential
concurrent accesses by using atomic accesses with relaxed memory order.