Ignore and clean configure~
Autoconf 2.71 saves an existing configure script to configure~ before
generating a new one. Althogh tools/autogen could delete configure~,
it may be helpful to not remove it so that it's possible to compare
the two versions of the script and make sure no unintended change
gets in.

So this commit adds configure~ to .gitignore and makes sure it gets
deleted during make clean.