remove the -force-tmc flag
-force-tmc was initially implemented by Frédéric Bour as a way to
implement "automatic tmc optimization" on top of the default behavior
which is "opt-in tmc optimization". We later decided to disable it,
because its usage was in tension with the goals we were setting for
the opt-in mode (in particular: have proper warnings and errors when
something that could be unexpected is going on).

It was kept for debugging purposes, and apparently found a way to
skean in the final merge for TMC (very probably my mistake), but in
a state where it is actually a no-op.