Config.mkexe should not include C flags that are used internally only
During the configure stage, two sets of C flags are distinguished:
those to be used to build programs during the compiler's build
(internal C flags) and those to be used when the compiler itself
builds user programs.

There was a mistake in the way the flags to use to build user programs
were computed, so that they ended up being built with internal flags.

One such flag was `-g`, which on some operating systems led to
the creation of spurious `.dSYM` files and directories.