check_well_founded: give expansions traces in error messages
Note: with the current way that well-foundedness is checked, there is
in fact little difference between the two different exceptions,
Recursive_abbrev and Cycle_in_def. (Recursive_abbrev sounds a bit more
confident in blaming the cycle precisely on a given definition, but
otherwise they are explained in the same way.) For now we suggest
keeping the two distinct errors instead of merging them, to see
whether they remain similar or grow more different with the different
algorithm in #11207.