An indirection for a memory-safe buffer implementation
In OCaml 4, the buffer module was using internally unsafe bytes accesses.
Those accesses relied on invariants directly dependent on the global runtime
lock to ensure safety even in case of concurrent uses of the same buffer.

Those invariants no longer holds in OCaml 5. As a consequence, it is no longer
possible to use unsafe accesses as liberally as in OCaml 4.

To reach a compromise between performance safety, this commit changes
the buffer implementation to fall back on safe functions in most cases,
but adds a fast path to performance sensitive functions whenever it is possible
to guarantee memory safety even in presence of data races (while offfering no
correctness guarantees).

More precisely, this commit changes the memory representation of buffer
by replacing the two mutable [length] and [buffer] fields by a single
mutable [inner] field which holds an immutable inner buffer.

This change makes it possible to enforece that for a inner buffer
[ib.length = Bytes.length ib.buffer].
Performance sensitive functions uses unsafe operation of this immutable
buffer whenever they don't need to mutate the inner buffer themselves.
Those "performance-sensitive" are currently:

- add_string
- add_char
- add_substring
- add_int8*
- add_int16*
- add_int32*
- add_int64*