Missing CAMLparam in win32's Unix.stat (#11737)
The `path` argument is used after a `caml_enter_blocking_section`
when the path does not exists (the path is used to build the
unix error exception).

Unfortunately, during the blocking section we may "yield" to
a thread that can trigger a garbage collection and move the
content of `path` elsewhere, triggering a segfault.

The CAMLparam is therefore needed in that case.

(cherry picked from commit 3d8fb96281c94d31299859e7a72b90ce0976254c)