Embed printer types in toplevels and debugger
The debugger and toplevels let users install printers for their data types.

To make sure the provided printers have the right type, both the debugger and
the toplevel need to know the internal OCaml representation of
the expected types.

So far, both the debugger and the toplevels used to extract the expected
representatioons from topdirs.cmi which they were reading at runtime.

This creates problems (e.g. when the compiler has not yet been installed)
and yields not so easy to read code.

In this commit, the type of the printers are defined in strings as they
would be written in OCaml in topdirs.ml. These strings are parsed
at runtime and the corresponding typing declarations are added to the
environment.