Generate the list of primitives using the preprocessor
The new code uses the preprocessor to expand possible macros in the
generation of the primitive prototype (for instance, some comparison
functions in runtime/floats.c). The preprocessor will however remove
the token CAMLprim, so we redefine in the call CAMLprim to
CAMLprim()=0, as a function-like macro will not be expanded if not
given any arguments.

The list of primitives is kept sorted to benefit from stability if it
doesn't change and functions just moved around.

A sed script was extracted in a separate file to simplify quoting and
escaping in the Makefile.

The caching mechanism of runtime/primitives is preserved: if the list
of primitives hasn't changed, this file is not modified.