toplevel: -{show,hide}-uninformative-sigs
see #10060

Currently, passing a type definition to the toplevel reprints it
unchanged:

    # type t = Foo;;
    type t = Foo
    # (* next prompt *)

When the new option -hide-uninformative-sigs is passed, the type
definition is not reprinted in this case, nothing is shown.

    # type t = Foo;;
    # (* next prompt *)

This hiding only happens when using the toplevel interactively,
not with #use directives that reprint everything.

The definition of 'uninformative signature' is not obvious, in
particular it does not correspond to a type-level vs. term-level
distinction:

- `external foo : ... = ...` has an uninformative signature
  (despite being term-level)
- `include List` has an informative signature

Our implementation is only an approximation (it both under- and
over-approximates):

- we consider that all modules and classes are informative
  (but arguably an unconstrained module that contains only type
  declarations has an uninformative signature)

- we consider that all module types and class types are uninformative
  (but arguably a module type that contains a signature inclusion
  has an informative signature)