Manual: add note about the change in behaviour of the runtime lock with multi-domain
It is not longer safe to assume that only one C bindings runs at a time,
even if that C binding never releases the OCaml runtime lock:
the OCaml program could have multiple domains, and the lock is now a
"domain lock".

Previously C bindings may have assumed that it is safe to
manipulate/initialize C global state as long as they do not release the
OCaml runtime lock, but that is no longer the case.
E.g. a C binding having a local `static` value is no longer safe, and
worth explicitly calling out in the manual!

Document in the parallelism section of the manual, and add a link to it from
the 'Interfacing with C chapter'.

Suggested-by: KC Sivaramakrishnan <kc@kcsrk.info>
Signed-off-by: Edwin Török <edwin.torok@cloud.com>