Manual: make new chapters on memory model and parallelism easier to discover
When I look at learning what is new in a release of OCaml I look at:
* Language extensions (an approx chronological order of language "diffs")
* The documentation in the .mli of newly introduced modules
* The Changes entry (which did contain a reference, albeit buried deep
  among other entries)

In this case the memory model and parallelism got their own chapters in
the main manual (which is well deserved), but it is also easy to miss
(it is not in the usual places I used to look for new information).

Improve this:

* mention the manual entry in the Atomic and Domain .mli docs.

* mention Effect handler manual entry in Effect.mli

* mention the new manual chapters in the top-level multicore changes
  entry where people are most likely to notice

The mention in the .mli refers to the chapter title, there isn't a convenient
and reliable way to add a link to html with ocamldoc/odoc at the moment
({%latex:\ref%} could be used for PDF).
This can be revisited in the future e.g. after some tooling changes.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>