Manual(Thread): mention domains
Mention that concurrency only applies within a domain (if there are other
domains this thread may run in parallel with those).
And add a link to Domain.spawn if the user is interested in parallelism instead
of concurrency.
(The Domain module's documentation contains a link to the manual chapter that
explains the distinction)

In other languages threads are used for parallelism too, so it is worth
pointing out that another mechanisms exists for that.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>