Reorganise how MKEXE_VIA_CC is built to make it correct for MSVC
MSVC cl command-line option `/link` allows to pass options to the linker
using the following scheme:
  `cl [cl arguments...] /link [link arguments...]`

In order to accomodate all our supported C compilers, this requires a
reshuffling of how the MKEXE_VIA_CC command line is built so that
`$oc_exe_ldflags` are indeed inserted after the `/link` for cl.
This commit also ensures that `/link` (`$mkexe_via_cc_ldflags_prefix`)
is inserted exactly once on the cl command line (rather than once per
argument, which is how flexlink expects these arguments).
Finally, this commit moves the usage of `mkexe_via_cc_extra_cmd` (which
is used only for the MSVC toolchain, to inject the manifest in the
executable) to the MKEXE_VIA_CC variable instead of hiding it into
`mkexe_via_cc_ldflags` where it doesn't belong.

For non-MSVC ports, this commit:
- removes the `$(OC_EXE_LDFLAGS)` Makefile variable as it is no longer
  used,
- and reorders the `${oc_exe_ldflags}` after the `$(CFLAGS)`.

Co-authored-by: Samuel Hym <samuel.hym@rustyne.lautre.net>