error messages: n-body signature avoidance
Some signature avoidance issues cannot be reduced to single illegal
include because the issue arise from the simultaneous erasure of more
than one type. For instance, in
```ocaml
module type t = sig
   include sig
     type t
   end
   include sig
     type s = t option
   end
   type t
   val x:s
   type s
end
```
neither the redefinition of `t` nor `s` are the only responsible for the
loss of a type name for the value `x`.

This commit takes care of reporting to the users all possibly
problematic redefinitions when we fail to anchor the type of a value to
an exported signature item.